+++ src/vmware/azext_vmware/_help.py
@@ -515,7 +515,7 @@ helps['vmware workload-network dhcp show'] = """
     short-summary: Get DHCP by ID in a private cloud workload network.
     examples:
     - name: Get DHCP by ID in a workload network.
-      text: az vmware workload-network dhcp show --resource-group group1 --private-cloud cloud1 --dhcp-id dhcp1
+      text: az vmware workload-network dhcp show --resource-group group1 --private-cloud cloud1 --dhcp dhcp1
 """
 
 helps['vmware workload-network dhcp relay'] = """
@@ -528,7 +528,7 @@ helps['vmware workload-network dhcp relay create'] = """
     short-summary: Create DHCP by ID in a private cloud workload network.
     examples:
     - name: Create DHCP by ID in a workload network.
-      text: az vmware workload-network dhcp relay create --resource-group group1 --private-cloud cloud1 --dhcp-id dhcp1 --display-name dhcpConfigurations1 --revision 1 --server-addresses 40.1.5.1/24
+      text: az vmware workload-network dhcp relay create --resource-group group1 --private-cloud cloud1 --dhcp dhcp1 --display-name dhcpConfigurations1 --revision 1 --server-addresses 40.1.5.1/24
 """
 
 helps['vmware workload-network dhcp relay delete'] = """
@@ -536,7 +536,7 @@ helps['vmware workload-network dhcp relay delete'] = """
     short-summary: Delete DHCP by ID in a private cloud workload network.
     examples:
     - name: Delete DHCP by ID in a workload network.
-      text: az vmware workload-network dhcp relay delete --resource-group group1 --private-cloud cloud1 --dhcp-id dhcp1
+      text: az vmware workload-network dhcp relay delete --resource-group group1 --private-cloud cloud1 --dhcp dhcp1
 """
 
 helps['vmware workload-network dhcp relay update'] = """
@@ -544,7 +544,7 @@ helps['vmware workload-network dhcp relay update'] = """
     short-summary: Update DHCP by ID in a private cloud workload network.
     examples:
     - name: Update DHCP by ID in a workload network.
-      text: az vmware workload-network dhcp relay update --resource-group group1 --private-cloud cloud1 --dhcp-id dhcp1 --display-name dhcpConfigurations1 --revision 1 --server-addresses 40.1.5.1/24
+      text: az vmware workload-network dhcp relay update --resource-group group1 --private-cloud cloud1 --dhcp dhcp1 --display-name dhcpConfigurations1 --revision 1 --server-addresses 40.1.5.1/24
 """
 
 helps['vmware workload-network dhcp server'] = """
@@ -557,7 +557,7 @@ helps['vmware workload-network dhcp server create'] = """
     short-summary: Create DHCP by ID in a private cloud workload network.
     examples:
     - name: Create DHCP by ID in a workload network.
-      text: az vmware workload-network dhcp server create --resource-group group1 --private-cloud cloud1 --dhcp-id dhcp1 --display-name dhcpConfigurations1 --revision 1 --server-address 40.1.5.1/24 --lease-time 86400
+      text: az vmware workload-network dhcp server create --resource-group group1 --private-cloud cloud1 --dhcp dhcp1 --display-name dhcpConfigurations1 --revision 1 --server-address 40.1.5.1/24 --lease-time 86400
 """
 
 helps['vmware workload-network dhcp server delete'] = """
@@ -565,7 +565,7 @@ helps['vmware workload-network dhcp server delete'] = """
     short-summary: Delete DHCP by ID in a private cloud workload network.
     examples:
     - name: Delete DHCP by ID in a workload network.
-      text: az vmware workload-network dhcp server delete --resource-group group1 --private-cloud cloud1 --dhcp-id dhcp1
+      text: az vmware workload-network dhcp server delete --resource-group group1 --private-cloud cloud1 --dhcp dhcp1
 """
 
 helps['vmware workload-network dhcp server update'] = """
@@ -573,7 +573,7 @@ helps['vmware workload-network dhcp server update'] = """
     short-summary: Update DHCP by ID in a private cloud workload network.
     examples:
     - name: Update DHCP by ID in a workload network.
-      text: az vmware workload-network dhcp server update --resource-group group1 --private-cloud cloud1 --dhcp-id dhcp1 --display-name dhcpConfigurations1 --revision 1 --server-address 40.1.5.1/24 --lease-time 86400
+      text: az vmware workload-network dhcp server update --resource-group group1 --private-cloud cloud1 --dhcp dhcp1 --display-name dhcpConfigurations1 --revision 1 --server-address 40.1.5.1/24 --lease-time 86400
 """
 
 
@@ -596,7 +596,7 @@ helps['vmware workload-network dns-service show'] = """
     short-summary: Get a DNS service by ID in a private cloud workload network.
     examples:
     - name: Get a DNS service by ID in a workload network.
-      text: az vmware workload-network dns-service show --resource-group group1 --private-cloud cloud1 --dns-service-id dnsService1
+      text: az vmware workload-network dns-service show --resource-group group1 --private-cloud cloud1 --dns-service dnsService1
 """
 
 helps['vmware workload-network dns-service create'] = """
@@ -604,7 +604,7 @@ helps['vmware workload-network dns-service create'] = """
     short-summary: Create a DNS service by ID in a private cloud workload network.
     examples:
     - name: Create a DNS service by ID in a workload network.
-      text: az vmware workload-network dns-service create --resource-group group1 --private-cloud cloud1 --dns-service-id dnsService1 --display-name dnsService1 --dns-service-ip 5.5.5.5 --default-dns-zone defaultDnsZone1 --fqdn-zones fqdnZone1 --log-level INFO --revision 1
+      text: az vmware workload-network dns-service create --resource-group group1 --private-cloud cloud1 --dns-service dnsService1 --display-name dnsService1 --dns-service-ip 5.5.5.5 --default-dns-zone defaultDnsZone1 --fqdn-zones fqdnZone1 --log-level INFO --revision 1
 """
 
 helps['vmware workload-network dns-service update'] = """
@@ -612,7 +612,7 @@ helps['vmware workload-network dns-service update'] = """
     short-summary: Update a DNS service by ID in a private cloud workload network.
     examples:
     - name: Update a DNS service by ID in a workload network.
-      text: az vmware workload-network dns-service update --resource-group group1 --private-cloud cloud1 --dns-service-id dnsService1 --display-name dnsService1 --dns-service-ip 5.5.5.5 --default-dns-zone defaultDnsZone1 --fqdn-zones fqdnZone1 --log-level INFO --revision 1
+      text: az vmware workload-network dns-service update --resource-group group1 --private-cloud cloud1 --dns-service dnsService1 --display-name dnsService1 --dns-service-ip 5.5.5.5 --default-dns-zone defaultDnsZone1 --fqdn-zones fqdnZone1 --log-level INFO --revision 1
 """
 
 helps['vmware workload-network dns-service delete'] = """
@@ -620,7 +620,7 @@ helps['vmware workload-network dns-service delete'] = """
     short-summary: Delete a DNS service by ID in a private cloud workload network.
     examples:
     - name: Delete a DNS service by ID in a workload network.
-      text: az vmware workload-network dns-service delete --resource-group group1 --private-cloud cloud1 --dns-service-id dnsService1
+      text: az vmware workload-network dns-service delete --resource-group group1 --private-cloud cloud1 --dns-service dnsService1
 """
 
 helps['vmware workload-network dns-zone'] = """
@@ -641,7 +641,7 @@ helps['vmware workload-network dns-zone show'] = """
     short-summary: Get a DNS zone by ID in a private cloud workload network.
     examples:
     - name: Get a DNS zone by ID in a workload network.
-      text: az vmware workload-network dns-zone show --resource-group group1 --private-cloud cloud1 --dns-zone-id dnsZone1
+      text: az vmware workload-network dns-zone show --resource-group group1 --private-cloud cloud1 --dns-zone dnsZone1
 """
 
 helps['vmware workload-network dns-zone create'] = """
@@ -649,7 +649,7 @@ helps['vmware workload-network dns-zone create'] = """
     short-summary: Create a DNS zone by ID in a private cloud workload network.
     examples:
     - name: Create a DNS zone by ID in a workload network.
-      text: az vmware workload-network dns-zone create --resource-group group1 --private-cloud cloud1 --dns-zone-id dnsZone1 --display-name dnsZone1 --domain domain1 --dns-server-ips 1.1.1.1 --source-ip 8.8.8.8 --dns-services 1 --revision 1
+      text: az vmware workload-network dns-zone create --resource-group group1 --private-cloud cloud1 --dns-zone dnsZone1 --display-name dnsZone1 --domain domain1 --dns-server-ips 1.1.1.1 --source-ip 8.8.8.8 --dns-services 1 --revision 1
 """
 
 helps['vmware workload-network dns-zone update'] = """
@@ -657,7 +657,7 @@ helps['vmware workload-network dns-zone update'] = """
     short-summary: Update a DNS zone by ID in a private cloud workload network.
     examples:
     - name: Update a DNS zone by ID in a workload network.
-      text: az vmware workload-network dns-zone update --resource-group group1 --private-cloud cloud1 --dns-zone-id dnsZone1 --display-name dnsZone1 --domain domain1 --dns-server-ips 1.1.1.1 --source-ip 8.8.8.8 --dns-services 1 --revision 1
+      text: az vmware workload-network dns-zone update --resource-group group1 --private-cloud cloud1 --dns-zone dnsZone1 --display-name dnsZone1 --domain domain1 --dns-server-ips 1.1.1.1 --source-ip 8.8.8.8 --dns-services 1 --revision 1
 """
 
 helps['vmware workload-network dns-zone delete'] = """
@@ -665,7 +665,7 @@ helps['vmware workload-network dns-zone delete'] = """
     short-summary: Delete a DNS zone by ID in a private cloud workload network.
     examples:
     - name: Delete a DNS zone by ID in a workload network.
-      text: az vmware workload-network dns-zone delete --resource-group group1 --private-cloud cloud1 --dns-zone-id dnsZone1
+      text: az vmware workload-network dns-zone delete --resource-group group1 --private-cloud cloud1 --dns-zone dnsZone1
 """
 
 helps['vmware workload-network port-mirroring'] = """
@@ -686,7 +686,7 @@ helps['vmware workload-network port-mirroring show'] = """
     short-summary: Get a port mirroring profile by ID in a private cloud workload network.
     examples:
     - name: Get a port mirroring profile by ID in a workload network.
-      text: az vmware workload-network port-mirroring show --resource-group group1 --private-cloud cloud1 --port-mirroring-id portMirroring1
+      text: az vmware workload-network port-mirroring show --resource-group group1 --private-cloud cloud1 --port-mirroring portMirroring1
 """
 
 helps['vmware workload-network port-mirroring create'] = """
@@ -694,7 +694,7 @@ helps['vmware workload-network port-mirroring create'] = """
     short-summary: Create a port mirroring profile by ID in a private cloud workload network.
     examples:
     - name: Create a port mirroring profile by ID in a workload network.
-      text: az vmware workload-network port-mirroring create --resource-group group1 --private-cloud cloud1 --port-mirroring-id portMirroring1 --display-name portMirroring1 --direction BIDIRECTIONAL --source vmGroup1 --destination vmGroup2 --revision 1
+      text: az vmware workload-network port-mirroring create --resource-group group1 --private-cloud cloud1 --port-mirroring portMirroring1 --display-name portMirroring1 --direction BIDIRECTIONAL --source vmGroup1 --destination vmGroup2 --revision 1
 """
 
 helps['vmware workload-network port-mirroring update'] = """
@@ -702,7 +702,7 @@ helps['vmware workload-network port-mirroring update'] = """
     short-summary: Update a port mirroring profile by ID in a private cloud workload network.
     examples:
     - name: Update a port mirroring profile by ID in a workload network.
-      text: az vmware workload-network port-mirroring update --resource-group group1 --private-cloud cloud1 --port-mirroring-id portMirroring1 --display-name portMirroring1 --direction BIDIRECTIONAL --source vmGroup1 --destination vmGroup2 --revision 1
+      text: az vmware workload-network port-mirroring update --resource-group group1 --private-cloud cloud1 --port-mirroring portMirroring1 --display-name portMirroring1 --direction BIDIRECTIONAL --source vmGroup1 --destination vmGroup2 --revision 1
 """
 
 helps['vmware workload-network port-mirroring delete'] = """
@@ -710,5 +710,174 @@ helps['vmware workload-network port-mirroring delete'] = """
     short-summary: Delete a port mirroring profile by ID in a private cloud workload network.
     examples:
     - name: Delete a port mirroring profile by ID in a workload network.
-      text: az vmware workload-network port-mirroring delete --resource-group group1 --private-cloud cloud1 --port-mirroring-id portMirroring1
+      text: az vmware workload-network port-mirroring delete --resource-group group1 --private-cloud cloud1 --port-mirroring portMirroring1
+"""
+
+helps['vmware workload-network segment'] = """
+    type: group
+    short-summary: Commands to manage a Segment workload network.
+"""
+
+helps['vmware workload-network segment list'] = """
+    type: command
+    short-summary: List of segments in a private cloud workload network.
+    examples:
+    - name: List of segments in a workload network.
+      text: az vmware workload-network segment list --resource-group group1 --private-cloud cloud1
+"""
+
+helps['vmware workload-network segment show'] = """
+    type: command
+    short-summary: Get a segment by ID in a private cloud workload network.
+    examples:
+    - name: Get a segment by ID in a workload network.
+      text: az vmware workload-network segment show --resource-group group1 --private-cloud cloud1 --segment segment1
+"""
+
+helps['vmware workload-network segment create'] = """
+    type: command
+    short-summary: Create a segment by ID in a private cloud workload network.
+    examples:
+    - name: Create a segment by ID in a workload network.
+      text: az vmware workload-network segment create --resource-group group1 --private-cloud cloud1 --segment segment1 --display-name segment1 --connected-gateway /infra/tier-1s/gateway --revision 1 --dhcp-ranges 40.20.0.0 40.20.0.1 --gateway-address 40.20.20.20/16 --port-name port1
+"""
+
+helps['vmware workload-network segment update'] = """
+    type: command
+    short-summary: Update a segment by ID in a private cloud workload network.
+    examples:
+    - name: Update a segment by ID in a workload network.
+      text: az vmware workload-network segment update --resource-group group1 --private-cloud cloud1 --segment segment1 --display-name segment1 --connected-gateway /infra/tier-1s/gateway --revision 1 --dhcp-ranges 40.20.0.0 40.20.0.1 --gateway-address 40.20.20.20/16 --port-name port1
+"""
+
+helps['vmware workload-network segment delete'] = """
+    type: command
+    short-summary: Delete a segment by ID in a private cloud workload network.
+    examples:
+    - name: Delete a segment by ID in a workload network.
+      text: az vmware workload-network segment delete --resource-group group1 --private-cloud cloud1 --segment segment1
+"""
+
+helps['vmware workload-network public-ip'] = """
+    type: group
+    short-summary: Commands to manage a Public-IP workload network.
+"""
+
+helps['vmware workload-network public-ip list'] = """
+    type: command
+    short-summary: List of Public IP Blocks in a private cloud workload network.
+    examples:
+    - name: List of Public IP Blocks in a workload network.
+      text: az vmware workload-network public-ip list --resource-group group1 --private-cloud cloud1
+"""
+
+helps['vmware workload-network public-ip show'] = """
+    type: command
+    short-summary: Get a Public IP Block by ID in a private cloud workload network.
+    examples:
+    - name: Get a Public IP Block by ID in a workload network.
+      text: az vmware workload-network public-ip show --resource-group group1 --private-cloud cloud1 --public-ip publicIP1
+"""
+
+helps['vmware workload-network public-ip create'] = """
+    type: command
+    short-summary: Create a Public IP Block by ID in a private cloud workload network.
+    examples:
+    - name: Create a Public IP Block by ID in a workload network.
+      text: az vmware workload-network public-ip create --resource-group group1 --private-cloud cloud1 --public-ip publicIP1 --display-name publicIP1 --number-of-public-ips 32
+"""
+
+helps['vmware workload-network public-ip delete'] = """
+    type: command
+    short-summary: Delete a Public IP Block by ID in a private cloud workload network.
+    examples:
+    - name: Delete a Public IP Block by ID in a workload network.
+      text: az vmware workload-network public-ip delete --resource-group group1 --private-cloud cloud1 --public-ip publicIP1
+"""
+
+helps['vmware workload-network vm-group'] = """
+    type: group
+    short-summary: Commands to manage a VM Group workload network.
+"""
+
+helps['vmware workload-network vm-group list'] = """
+    type: command
+    short-summary: List of VM Groups in a private cloud workload network.
+    examples:
+    - name: List of VM Groups in a workload network.
+      text: az vmware workload-network vm-group list --resource-group group1 --private-cloud cloud1
+"""
+
+helps['vmware workload-network vm-group show'] = """
+    type: command
+    short-summary: Get a VM Group by ID in a private cloud workload network.
+    examples:
+    - name: Get a VM Group by ID in a workload network.
+      text: az vmware workload-network vm-group show --resource-group group1 --private-cloud cloud1 --vm-group vmGroup1
+"""
+
+helps['vmware workload-network vm-group create'] = """
+    type: command
+    short-summary: Create a VM Group by ID in a private cloud workload network.
+    examples:
+    - name: Create a VM Group by ID in a workload network.
+      text: az vmware workload-network vm-group create --resource-group group1 --private-cloud cloud1 --vm-group vmGroup1 --display-name vmGroup1 --members 564d43da-fefc-2a3b-1d92-42855622fa50 --revision 1
+"""
+
+helps['vmware workload-network vm-group update'] = """
+    type: command
+    short-summary: Update a VM Group by ID in a private cloud workload network.
+    examples:
+    - name: Update a VM Group by ID in a workload network.
+      text: az vmware workload-network vm-group update --resource-group group1 --private-cloud cloud1 --vm-group vmGroup1 --display-name vmGroup1 --members 564d43da-fefc-2a3b-1d92-42855622fa50 --revision 1
+"""
+
+helps['vmware workload-network vm-group delete'] = """
+    type: command
+    short-summary: Delete a VM Group by ID in a private cloud workload network.
+    examples:
+    - name: Delete a VM Group by ID in a private cloud workload network.
+      text: az vmware workload-network vm-group delete --resource-group group1 --private-cloud cloud1 --vm-group vmGroup1
+"""
+
+helps['vmware workload-network vm'] = """
+    type: group
+    short-summary: Commands to manage a Virtual Machine workload network.
+"""
+
+helps['vmware workload-network vm list'] = """
+    type: command
+    short-summary: List of Virtual Machines in a private cloud workload network.
+    examples:
+    - name: List of Virtual Machines in a workload network.
+      text: az vmware workload-network vm list --resource-group group1 --private-cloud cloud1
+"""
+
+helps['vmware workload-network vm show'] = """
+    type: command
+    short-summary: Get a Virtual Machines by ID in a private cloud workload network.
+    examples:
+    - name: Get a Virtual Machines by ID in a workload network.
+      text: az vmware workload-network vm show --resource-group group1 --private-cloud cloud1 --virtual-machine vm1
+"""
+
+helps['vmware workload-network gateway'] = """
+    type: group
+    short-summary: Commands to manage a Gateway workload network.
+"""
+
+helps['vmware workload-network gateway list'] = """
+    type: command
+    short-summary: List of Gateways in a private cloud workload network.
+    examples:
+    - name: List of Gateways in a workload network.
+      text: az vmware workload-network gateway list --resource-group group1 --private-cloud cloud1
+"""
+
+helps['vmware workload-network gateway show'] = """
+    type: command
+    short-summary: Get a Gateway by ID in a private cloud workload network.
+    examples:
+    - name: Get a Gateway by ID in a workload network.
+      text: az vmware workload-network gateway show --resource-group group1 --private-cloud cloud1 --gateway gateway1
 """